Recoding block page
